Don't rely on GLM for Vec3 functions
