+++ simulant/math/vec3.cpp
@@ -1,0 +1,78 @@
+#include "../types.h"
+#include "../utils/random.h"
+
+#include "vec3.h"
+
+namespace smlt {
+
+Vec3::Vec3(const Vec2 &v2, float z):
+    x(v2.x), y(v2.y), z(z) {
+
+}
+
+Vec3 Vec3::rotated_by(const Quaternion &q) const {
+    return q * (*this);
+}
+
+Vec3 Vec3::rotated_by(const Mat3 &rot) const {
+    return rot.transform_vector(*this);
+}
+
+Vec3 operator*(float lhs, const Vec3& rhs) {
+    smlt::Vec3 result = rhs;
+    result *= lhs;
+    return result;
+}
+
+Vec3 operator/(float lhs, const Vec3& rhs) {
+    smlt::Vec3 result = rhs;
+    result /= lhs;
+    return result;
+}
+
+Vec3 Vec3::random_deviant(const Degrees& angle, const Vec3 up) const {
+    //Lovingly adapted from ogre
+    Vec3 new_up = (up == Vec3()) ? perpendicular() : up;
+
+    Quaternion q(*this, Radians(random_gen::random_float(0, 1) * (PI * 2.0)));
+
+    new_up = new_up * q;
+
+    q = Quaternion(new_up, angle);
+
+    return *this * q;
+
+}
+
+Vec3 Vec3::operator*(const Quaternion &rhs) const {
+    return rhs.rotate_vector(*this);
+}
+
+Vec3 Vec3::rotated_by(const Mat4 &rot) const {
+    auto tmp = rot * Vec4(*this, 0);
+    return Vec3(tmp.x, tmp.y, tmp.z);
+}
+
+Vec3 Vec3::transformed_by(const Mat4 &trans) const {
+    auto tmp = trans * Vec4(*this, 1);
+    return Vec3(tmp.x, tmp.y, tmp.z);
+}
+
+Vec3 Vec3::perpendicular() const {
+    //Lovingly adapted from Ogre
+    static const float square_zero = (float)(1e-06 * 1e-06);
+    Vec3 perp = this->cross(Vec3(1, 0, 0));
+
+    if(perp.length_squared() < square_zero) {
+        //This vector is the X-axis, so use another
+        perp = this->cross(Vec3(0, 1, 0));
+    }
+    return perp.normalized();
+}
+
+smlt::Vec3 operator-(const smlt::Vec3& vec) {
+    return smlt::Vec3(-vec.x, -vec.y, -vec.z);
+}
+
+
+}
